Created Watched and WatchList buttons

+++ src/main/java/com/movtrack/List/DB/ListRepository.java
@@ -8,6 +8,6 @@ import java.util.List;
 
 public interface ListRepository extends CrudRepository<Media, Integer> {
     List<Media> findByListType(ListType listType);
-    List<Media> findByMediaID(long mediaID);
+    List<Media> findByMediaID(int mediaID);
     List<Media> findByMediaType(String mediaType);
 }
